+++ packages/toasts/src/components/DefaultToast.vue
@@ -13,10 +13,8 @@
 </template>
 
 <script lang="ts" setup>
+import { Icon, ICON_NAMES } from '@tokene/ui-kit'
 import { computed } from 'vue'
-
-import { Icon } from '@/common'
-import { ICON_NAMES } from '@/enums'
 
 const props = withDefaults(
   defineProps<{
@@ -34,8 +32,6 @@ const toastIcon = computed(() => props.iconName || ICON_NAMES.checkFilled)
 </script>
 
 <style lang="scss">
-@import 'vue-toastification/src/scss/index';
-
 .Vue-Toastification__toast {
   &--success {
     background: var(--success-dark);