+++ packages/toasts/src/composables/use-notifications.ts
@@ -1,0 +1,121 @@
+import { i18next } from '@tokene/toolkit'
+import { ICON_NAMES } from '@tokene/ui-kit'
+import { isObject } from 'lodash-es'
+import { TYPE, useToast } from 'vue-toastification'
+
+import { DefaultToast, TransactionToast } from '@/components'
+import {
+  CommonNotificationTypes,
+  NotificationObjectPayload,
+  NotificationTxType,
+} from '@/types'
+
+const MINUTE = 60 * 1000
+
+export const useNotifications = () => {
+  const toast = useToast()
+
+  // FIXME
+  const { t } = i18next
+
+  const defaultTitles = {
+    [TYPE.SUCCESS]: t('notification.default-title-success'),
+    [TYPE.ERROR]: t('notification.default-title-error'),
+    [TYPE.WARNING]: t('notification.default-title-warning'),
+    [TYPE.INFO]: t('notification.default-title-info'),
+    [TYPE.DEFAULT]: t('notification.default-title-default'),
+  }
+  const defaultMessages = {
+    [TYPE.DEFAULT]: t('notification.default-message-default'),
+    [TYPE.INFO]: t('notification.default-message-info'),
+    [TYPE.SUCCESS]: t('notification.default-message-success'),
+    [TYPE.ERROR]: t('notification.default-message-error'),
+    [TYPE.WARNING]: t('notification.default-message-warning'),
+  }
+  const defaultIconNames = {
+    [TYPE.DEFAULT]: ICON_NAMES.exclamationCircleFilled,
+    [TYPE.INFO]: ICON_NAMES.exclamationCircleFilled,
+    [TYPE.SUCCESS]: ICON_NAMES.checkFilled,
+    [TYPE.ERROR]: ICON_NAMES.exclamationCircleFilled,
+    [TYPE.WARNING]: ICON_NAMES.triangleFilled,
+  }
+
+  const showToast = (
+    messageType = 'default' as CommonNotificationTypes,
+    payload?: string | NotificationObjectPayload,
+  ) => {
+    let title = ''
+    let message = ''
+    let iconName: ICON_NAMES | undefined
+
+    if (isObject(payload)) {
+      const msgPayload = payload as NotificationObjectPayload
+
+      title = msgPayload.title || ''
+      message = msgPayload.message
+      iconName = msgPayload.iconName
+    } else if (payload) {
+      message = payload as string
+    } else {
+      message = defaultMessages[messageType]
+    }
+
+    if (!title) {
+      title = defaultTitles[messageType]
+    }
+    if (!iconName) {
+      iconName = defaultIconNames[messageType]
+    }
+
+    return toast(
+      {
+        component: DefaultToast,
+        props: {
+          ...(title && { title }),
+          message,
+          ...(iconName && { iconName }),
+        },
+      },
+      {
+        icon: false,
+        type: {
+          default: TYPE.DEFAULT,
+          info: TYPE.INFO,
+          success: TYPE.SUCCESS,
+          error: TYPE.ERROR,
+          warning: TYPE.WARNING,
+        }[messageType],
+        timeout: MINUTE / 2,
+        closeOnClick: false,
+      },
+    )
+  }
+
+  const showTxToast = (type: NotificationTxType, link: string) => {
+    return toast(
+      {
+        component: TransactionToast,
+        props: {
+          type: type,
+          link: link,
+        },
+      },
+      {
+        icon: false,
+        type: {
+          pending: TYPE.INFO,
+          success: TYPE.SUCCESS,
+          error: TYPE.ERROR,
+        }[type],
+        timeout: MINUTE / 2,
+        closeOnClick: false,
+      },
+    )
+  }
+
+  const removeToast = (id: string | number) => toast.dismiss(id)
+
+  return { showToast, showTxToast, removeToast }
+}
+
+export * from 'vue-toastification'