+++ packages/toolkit/src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './config'
 export * from './event-bus'
-export * from './notifications'