+++ packages/toolkit/src/types/notifications.ts
@@ -1,7 +1,0 @@
-export type NotificationObjectPayload = {
-  title?: string
-  message: string
-  iconName?: string // FIXME
-}
-
-export type NotificationPayload = string | NotificationObjectPayload