+++ packages/ui-kit/src/common/Accordion.vue
@@ -33,7 +33,7 @@
 <script lang="ts" setup>
 import { onClickOutside } from '@vueuse/core'
 import { onMounted, ref } from 'vue'
-import { useRouter } from 'vue-router'
+import { onBeforeRouteUpdate } from 'vue-router'
 
 const props = withDefaults(
   defineProps<{
@@ -48,9 +48,8 @@ const props = withDefaults(
 
 const rootEl = ref<HTMLElement | null>(null)
 const isAccordionOpen = ref(props.isOpenedByDefault)
-const router = useRouter()
 
-router.afterEach(() => {
+onBeforeRouteUpdate(() => {
   closeAccordion()
 })